Fixed minor issue using get methods to delete items in admin.  Found by Janek Vind (waraxe)

+++ include/admin/customprofile.php
@@ -33,14 +33,10 @@
         if(!isset($_POST['html_disabled']))
             $_POST['html_disabled']=0;
 
-        if(!isset($_POST['show_in_admin']))
-            $_POST['show_in_admin']=0;
-
         if($_POST['curr'] == 'NEW') {
             // checking names of existing fields
             foreach($PHORUM['PROFILE_FIELDS'] as $id => $profile_field) {
-                if ($id === 'num_fields') continue;
-                if($profile_field['name'] == $_POST['string']) {
+                if($id !== 'num_fields' && $profile_field['name'] == $_POST['string']) {
                     $exists_already = true;
                     break;
                 }
@@ -66,34 +62,36 @@
                 $PHORUM["PROFILE_FIELDS"][$_POST["curr"]]['name']=$_POST["string"];
                 $PHORUM["PROFILE_FIELDS"][$_POST["curr"]]['length']=$_POST['length'];
                 $PHORUM["PROFILE_FIELDS"][$_POST["curr"]]['html_disabled']=$_POST['html_disabled'];
-                $PHORUM["PROFILE_FIELDS"][$_POST["curr"]]['show_in_admin']=$_POST['show_in_admin'];
             } else { // adding a new field
                 $PHORUM['PROFILE_FIELDS']["num_fields"]++;
                 $PHORUM["PROFILE_FIELDS"][$PHORUM['PROFILE_FIELDS']["num_fields"]]=array();
                 $PHORUM["PROFILE_FIELDS"][$PHORUM['PROFILE_FIELDS']["num_fields"]]['name']=$_POST["string"];
                 $PHORUM["PROFILE_FIELDS"][$PHORUM['PROFILE_FIELDS']["num_fields"]]['length']=$_POST['length'];
                 $PHORUM["PROFILE_FIELDS"][$PHORUM['PROFILE_FIELDS']["num_fields"]]['html_disabled']=$_POST['html_disabled'];
-                $PHORUM["PROFILE_FIELDS"][$PHORUM['PROFILE_FIELDS']["num_fields"]]['show_in_admin']=$_POST['show_in_admin'];
             }
 
             if(!phorum_db_update_settings(array("PROFILE_FIELDS"=>$PHORUM["PROFILE_FIELDS"]))){
                 $error="Database error while updating settings.";
             } else {
-                echo "Profile Field Updated<br />";
+                phorum_admin_okmsg("Profile Field Updated");
             }
 
         }
 
     }
 
-    if(isset($_GET["curr"])){
-        if(isset($_GET["delete"])){
-            unset($PHORUM["PROFILE_FIELDS"][$_GET["curr"]]);
-            phorum_db_update_settings(array("PROFILE_FIELDS"=>$PHORUM["PROFILE_FIELDS"]));
-            echo "Profile Field Deleted<br />";
+    if(isset($_POST["curr"]) && isset($_POST["delete"]) && $_POST["confirm"]=="Yes"){
+        $_POST["curr"] = (int)$_POST["curr"];
+        unset($PHORUM["PROFILE_FIELDS"][$_POST["curr"]]);
+        if(!phorum_db_update_settings(array("PROFILE_FIELDS"=>$PHORUM["PROFILE_FIELDS"]))){
+            $error="Database error while updating settings.";
         } else {
-            $curr = $_GET["curr"];
+            phorum_admin_okmsg("Profile Field Deleted");
         }
+    }
+
+    if(isset($_GET["curr"])){
+        $curr = (int)$_GET["curr"];
     }
 
 
@@ -101,7 +99,6 @@
         $string=$PHORUM["PROFILE_FIELDS"][$curr]['name'];
         $length=$PHORUM["PROFILE_FIELDS"][$curr]['length'];
         $html_disabled=$PHORUM["PROFILE_FIELDS"][$curr]['html_disabled'];
-        $show_in_admin=$PHORUM["PROFILE_FIELDS"][$curr]['show_in_admin'];
         $title="Edit Profile Field";
         $submit="Update";
     } else {
@@ -110,67 +107,82 @@
         $submit="Add";
         $length=255;
         $html_disabled=1;
-        $show_in_admin=1;
     }
 
     if($error){
         phorum_admin_error($error);
     }
 
-    include_once "./include/admin/PhorumInputForm.php";
+    if($_GET["curr"] && $_GET["delete"]){
 
-    $frm = new PhorumInputForm ("", "post", $submit);
+        ?>
 
-    $frm->hidden("module", "customprofile");
+        <div class="PhorumInfoMessage">
+            Are you sure you want to delete this entry?
+            <form action="<?php echo $_SERVER["PHP_SELF"]; ?>" method="post">
+                <input type="hidden" name="module" value="<?php echo $module; ?>" />
+                <input type="hidden" name="curr" value="<?php echo $_GET['curr']; ?>" />
+                <input type="hidden" name="delete" value="1" />
+                <input type="submit" name="confirm" value="Yes" />&nbsp;<input type="submit" name="confirm" value="No" />
+            </form>
+        </div>
 
-    $frm->hidden("curr", "$curr");
+        <?php
 
-    $frm->addbreak($title);
+    } else {
 
-    $frm->addrow("Field Name", $frm->text_box("string", $string, 50));
-    $frm->addrow("Field Length (Max. 65000)", $frm->text_box("length", $length, 50));
-    $frm->addrow("Disable HTML", $frm->checkbox("html_disabled",1,"Yes",$html_disabled));
-    $frm->addrow("Show that field in the user-admin", $frm->checkbox("show_in_admin",1,"Yes",$show_in_admin));
 
-    $frm->show();
+        include_once "./include/admin/PhorumInputForm.php";
 
-    echo "This will only add the field to the list of allowed fields.  You will need to edit the register and profile templates to actually allow users to use the fields.  Use the name you enter here as the name property of the HTML form element.";
+        $frm =& new PhorumInputForm ("", "post", $submit);
 
-    if($curr=="NEW"){
+        $frm->hidden("module", "customprofile");
 
-        echo "<hr class=\"PhorumAdminHR\" />";
-        if(isset($PHORUM['PROFILE_FIELDS']["num_fields"]))
-            unset($PHORUM['PROFILE_FIELDS']["num_fields"]);
+        $frm->hidden("curr", "$curr");
 
-        if(count($PHORUM["PROFILE_FIELDS"])){
+        $frm->addbreak($title);
 
-            echo "<table border=\"0\" cellspacing=\"1\" cellpadding=\"0\" class=\"PhorumAdminTable\" width=\"100%\">\n";
-            echo "<tr>\n";
-            echo "    <td class=\"PhorumAdminTableHead\">Field</td>\n";
-            echo "    <td class=\"PhorumAdminTableHead\">Length</td>\n";
-            echo "    <td class=\"PhorumAdminTableHead\">HTML disabled</td>\n";
-            echo "    <td class=\"PhorumAdminTableHead\">Show in admin</td>\n";
-            echo "    <td class=\"PhorumAdminTableHead\">&nbsp;</td>\n";
-            echo "</tr>\n";
+        $frm->addrow("Field Name", $frm->text_box("string", $string, 50));
+        $frm->addrow("Field Length (Max. 65000)", $frm->text_box("length", $length, 50));
+        $frm->addrow("Disable HTML", $frm->checkbox("html_disabled",1,"Yes",$html_disabled));
 
-            foreach($PHORUM["PROFILE_FIELDS"] as $key => $item){
+        $frm->show();
+
+        echo "This will only add the field to the list of allowed fields.  You will need to edit the register and profile templates to actually allow users to use the fields.  Use the name you enter here as the name property of the HTML form element.";
+
+        if($curr=="NEW"){
+
+            echo "<hr class=\"PhorumAdminHR\" />";
+            if(isset($PHORUM['PROFILE_FIELDS']["num_fields"]))
+                unset($PHORUM['PROFILE_FIELDS']["num_fields"]);
+
+            if(count($PHORUM["PROFILE_FIELDS"])){
+
+                echo "<table border=\"0\" cellspacing=\"1\" cellpadding=\"0\" class=\"PhorumAdminTable\" width=\"100%\">\n";
                 echo "<tr>\n";
-                echo "    <td class=\"PhorumAdminTableRow\">".$item['name']."</td>\n";
-                echo "    <td class=\"PhorumAdminTableRow\">".$item['length']."</td>\n";
-                echo "    <td class=\"PhorumAdminTableRow\">".($item['html_disabled']?"Yes":"No")."</td>\n";
-                echo "    <td class=\"PhorumAdminTableRow\">".($item['show_in_admin']?"Yes":"No")."</td>\n";
-                echo "    <td class=\"PhorumAdminTableRow\"><a href=\"$_SERVER[PHP_SELF]?module=customprofile&curr=$key&edit=1\">Edit</a>&nbsp;&#149;&nbsp;<a href=\"$_SERVER[PHP_SELF]?module=customprofile&curr=$key&delete=1\">Delete</a></td>\n";
+                echo "    <td class=\"PhorumAdminTableHead\">Field</td>\n";
+                echo "    <td class=\"PhorumAdminTableHead\">Length</td>\n";
+                echo "    <td class=\"PhorumAdminTableHead\">HTML disabled</td>\n";
+                echo "    <td class=\"PhorumAdminTableHead\">&nbsp;</td>\n";
                 echo "</tr>\n";
+
+                foreach($PHORUM["PROFILE_FIELDS"] as $key => $item){
+                    echo "<tr>\n";
+                    echo "    <td class=\"PhorumAdminTableRow\">".$item['name']."</td>\n";
+                    echo "    <td class=\"PhorumAdminTableRow\">".$item['length']."</td>\n";
+                    echo "    <td class=\"PhorumAdminTableRow\">".($item['html_disabled']?"Yes":"No")."</td>\n";
+                    echo "    <td class=\"PhorumAdminTableRow\"><a href=\"$_SERVER[PHP_SELF]?module=customprofile&curr=$key&?edit=1\">Edit</a>&nbsp;&#149;&nbsp;<a href=\"$_SERVER[PHP_SELF]?module=customprofile&curr=$key&delete=1\">Delete</a></td>\n";
+                    echo "</tr>\n";
+                }
+
+                echo "</table>\n";
+
+            } else {
+
+                echo "No custom fields currently allowed.";
+
             }
 
-            echo "</table>\n";
-
-        } else {
-
-            echo "No custom fields currently allowed.";
-
         }
-
     }
-
 ?>
